Call init_logs() function to add logging handlers. Move init_logs() into initialization vs logging.
